Useless files removed
Simplified the contents of the folder so that only files that I am
working on remain in the directory. I am also not using the
'commit -m' flag for this message since I am now comfortable
with basic vim. I keep the line length to 75 characters as
advised by the higher authorities.